Punishment SQL Foreign key violation resolved

+++ src/main/java/com/prison/project/service/punishment/DeletePunishmentService.java
@@ -1,9 +1,10 @@
 package com.prison.project.service.punishment;
+
 import com.prison.project.repository.PunishmentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
 import javax.transaction.Transactional;
+import java.util.List;
 
 @Transactional
 @Service
@@ -12,7 +13,16 @@ public class DeletePunishmentService {
 
     private final PunishmentRepository punishmentRepository;
 
-    public void deletePunishment(Long id){
-        punishmentRepository.deleteById(id);
+    public void deletePunishment(Long id) {
+
+        List<Long> punishmentList = punishmentRepository.getPrisonerPunishment(id);
+        if (punishmentList.isEmpty()) {
+            punishmentRepository.deleteById(id);
+        }
     }
+
+    public List<Long> getPunishmentPrisoner(Long id) {
+        return punishmentRepository.getPrisonerPunishment(id);
+    }
+
 }
